+++ src/controllers/ElementsController.php
@@ -19,6 +19,7 @@ use craft\errors\InvalidElementException;
 use craft\errors\InvalidTypeException;
 use craft\errors\UnsupportedSiteException;
 use craft\events\DefineElementEditorHtmlEvent;
+use craft\events\DraftEvent;
 use craft\helpers\ArrayHelper;
 use craft\helpers\Component;
 use craft\helpers\Cp;
@@ -29,6 +30,7 @@ use craft\helpers\UrlHelper;
 use craft\i18n\Locale;
 use craft\models\ElementActivity;
 use craft\models\FieldLayoutForm;
+use craft\services\Drafts;
 use craft\web\Controller;
 use craft\web\CpScreenResponseBehavior;
 use craft\web\UrlManager;
@@ -78,7 +80,6 @@ class ElementsController extends Controller
     private ?string $_draftName = null;
     private ?string $_notes = null;
     private string $_fieldsLocation;
-    private ?array $_dirtyFields = null;
     private bool $_provisional;
     private bool $_dropProvisional;
     private bool $_addAnother;
@@ -115,7 +116,6 @@ class ElementsController extends Controller
         $this->_draftName = $this->_param('draftName');
         $this->_notes = $this->_param('notes');
         $this->_fieldsLocation = $this->_param('fieldsLocation') ?? 'fields';
-        $this->_dirtyFields = $this->_param('dirtyFields');
         $this->_provisional = (bool)$this->_param('provisional');
         $this->_dropProvisional = (bool)$this->_param('dropProvisional');
         $this->_addAnother = (bool)$this->_param('addAnother');
@@ -238,7 +238,7 @@ class ElementsController extends Controller
 
         if ($element === null) {
             /** @var Element|DraftBehavior|RevisionBehavior|Response|null $element */
-            $element = $this->_element($elementId, null, true, $strictSite);
+            $element = $this->_element($elementId, checkForProvisionalDraft: true, strictSite: $strictSite);
 
             if ($element instanceof Response) {
                 return $element;
@@ -525,7 +525,7 @@ class ElementsController extends Controller
         $this->requireCpRequest();
 
         /** @var Element|DraftBehavior|RevisionBehavior|Response|null $element */
-        $element = $this->_element($elementId, null, false);
+        $element = $this->_element($elementId);
 
         if (!$element) {
             throw new BadRequestHttpException('No element was identified by the request.');
@@ -1363,7 +1363,7 @@ JS, [
         $this->requirePostRequest();
 
         /** @var Element|null $element */
-        $element = $this->_element(provisional: true);
+        $element = $this->_element(checkForProvisionalDraft: true);
 
         if (!$element || $element->getIsRevision()) {
             throw new BadRequestHttpException('No element was identified by the request.');
@@ -1441,11 +1441,21 @@ JS, [
             }
         }
 
-        return Craft::$app->getDb()->transaction(function() use ($element, $user, $elementsService): ?Response {
+        // Keep track of all newly-created draft IDs
+        $draftElementIds = [];
+        $draftsService = Craft::$app->getDrafts();
+        $draftsService->on(Drafts::EVENT_AFTER_CREATE_DRAFT, function(DraftEvent $event) use (&$draftElementIds) {
+            $draftElementIds[$event->canonical->id] = $event->draft->id;
+        });
+
+        $db = Craft::$app->getDb();
+        $transaction = $db->beginTransaction();
+
+        try {
             // Are we creating the draft here?
             if (!$element->getIsDraft()) {
                 /** @var Element|DraftBehavior $element */
-                $draft = Craft::$app->getDrafts()->createDraft($element, $user->id, null, null, [], $this->_provisional);
+                $draft = $draftsService->createDraft($element, $user->id, null, null, [], $this->_provisional);
                 $draft->setCanonical($element);
                 $element = $this->element = $draft;
             }
@@ -1464,46 +1474,108 @@ JS, [
             $element->setScenario(Element::SCENARIO_ESSENTIALS);
 
             if (!$elementsService->saveElement($element)) {
+                $transaction->rollBack();
                 return $this->_asFailure($element, Craft::t('app', 'Couldn’t save {type}.', [
                     'type' => Craft::t('app', 'draft'),
                 ]));
             }
 
-            $elementsService->trackActivity($element, ElementActivity::TYPE_SAVE);
+            $transaction->commit();
+        } catch (Throwable $e) {
+            $transaction->rollBack();
+            throw $e;
+        }
 
-            $creator = $element->getCreator();
+        $elementsService->trackActivity($element, ElementActivity::TYPE_SAVE);
 
-            $data = [
-                'canonicalId' => $element->getCanonicalId(),
-                'elementId' => $element->id,
-                'draftId' => $element->draftId,
-                'timestamp' => Craft::$app->getFormatter()->asTimestamp($element->dateUpdated, 'short', true),
-                'creator' => $creator?->getName(),
-                'draftName' => $element->draftName,
-                'draftNotes' => $element->draftNotes,
-                'modifiedAttributes' => $element->getModifiedAttributes(),
+        $creator = $element->getCreator();
+
+        $data = [
+            'canonicalId' => $element->getCanonicalId(),
+            'elementId' => $element->id,
+            'draftId' => $element->draftId,
+            'timestamp' => Craft::$app->getFormatter()->asTimestamp($element->dateUpdated, 'short', true),
+            'creator' => $creator?->getName(),
+            'draftName' => $element->draftName,
+            'draftNotes' => $element->draftNotes,
+            'modifiedAttributes' => $element->getModifiedAttributes(),
+            'draftElementIds' => $draftElementIds,
+        ];
+
+        if ($this->request->getIsCpRequest()) {
+            [$docTitle, $title] = $this->_editElementTitles($element);
+            $data += $this->_fieldLayoutData($element);
+            $data += [
+                'docTitle' => $docTitle,
+                'title' => $title,
+                'previewTargets' => $element->getPreviewTargets(),
+                'initialDeltaValues' => Craft::$app->getView()->getInitialDeltaValues(),
+                'updatedTimestamp' => $element->dateUpdated->getTimestamp(),
+                'canonicalUpdatedTimestamp' => $element->getCanonical()->dateUpdated->getTimestamp(),
             ];
+        }
 
-            if ($this->request->getIsCpRequest()) {
-                [$docTitle, $title] = $this->_editElementTitles($element);
-                $data += $this->_fieldLayoutData($element);
-                $data += [
-                    'docTitle' => $docTitle,
-                    'title' => $title,
-                    'previewTargets' => $element->getPreviewTargets(),
-                    'initialDeltaValues' => Craft::$app->getView()->getInitialDeltaValues(),
-                    'updatedTimestamp' => $element->dateUpdated->getTimestamp(),
-                    'canonicalUpdatedTimestamp' => $element->getCanonical()->dateUpdated->getTimestamp(),
-                ];
-            }
+        // Make sure the user is authorized to preview the draft
+        Craft::$app->getSession()->authorize("previewDraft:$element->draftId");
 
-            // Make sure the user is authorized to preview the draft
-            Craft::$app->getSession()->authorize("previewDraft:$element->draftId");
+        return $this->_asSuccess(Craft::t('app', '{type} saved.', [
+            'type' => Craft::t('app', 'Draft'),
+        ]), $element, $data, true);
+    }
 
-            return $this->_asSuccess(Craft::t('app', '{type} saved.', [
-                'type' => Craft::t('app', 'Draft'),
-            ]), $element, $data, true);
-        });
+    /**
+     * Ensures that a provisional draft exists for the element, unless it’s already a draft.
+     *
+     * @return Response
+     * @since 5.0.0
+     */
+    public function actionEnsureDraft(): Response
+    {
+        $this->requirePostRequest();
+
+        /** @var Element|DraftBehavior|null $element */
+        $element = $this->_element(checkForProvisionalDraft: true);
+
+        if (!$element || $element->getIsRevision()) {
+            throw new BadRequestHttpException('No element was identified by the request.');
+        }
+
+        if ($element->getIsDraft()) {
+            return $this->asSuccess(data: [
+                'elementId' => $element->id,
+            ]);
+        }
+
+        $elementsService = Craft::$app->getElements();
+        $user = static::currentUser();
+
+        if (!$elementsService->canCreateDrafts($element, $user)) {
+            throw new ForbiddenHttpException('User not authorized to create drafts for this element.');
+        }
+
+        $this->element = $element;
+
+        // Make sure a provisional draft doesn't already exist for this element/user combo
+        $provisionalId = $element::find()
+            ->provisionalDrafts()
+            ->draftOf($element->id)
+            ->draftCreator($user->id)
+            ->site('*')
+            ->status(null)
+            ->ids()[0] ?? null;
+
+        if ($provisionalId) {
+            return $this->asSuccess(data: [
+                'elementId' => $provisionalId,
+            ]);
+        }
+
+        /** @var Element|DraftBehavior $element */
+        $draft = Craft::$app->getDrafts()->createDraft($element, $user->id, provisional: true);
+
+        return $this->asSuccess(data: [
+            'elementId' => $draft->id,
+        ]);
     }
 
     /**
@@ -1882,14 +1954,18 @@ JS, [
      *
      * @param int|null $elementId
      * @param string|null $elementUid
-     * @param bool|null $provisional
+     * @param bool $checkForProvisionalDraft
      * @param bool $strictSite
      * @return ElementInterface|Response|null
      * @throws BadRequestHttpException
      * @throws ForbiddenHttpException
      */
-    private function _element(?int $elementId = null, ?string $elementUid = null, ?bool $provisional = null, bool $strictSite = true): ElementInterface|Response|null
-    {
+    private function _element(
+        ?int $elementId = null,
+        ?string $elementUid = null,
+        bool $checkForProvisionalDraft = false,
+        bool $strictSite = true,
+    ): ElementInterface|Response|null {
         $elementId = $elementId ?? $this->_elementId;
         $elementUid = $elementUid ?? $this->_elementUid;
 
@@ -1957,7 +2033,7 @@ JS, [
         } elseif ($elementId || $elementUid) {
             if ($elementId) {
                 // First check for a provisional draft, if we're open to it
-                if ($provisional) {
+                if ($checkForProvisionalDraft) {
                     $element = $elementType::find()
                         ->provisionalDrafts()
                         ->draftOf($elementId)
@@ -2127,11 +2203,6 @@ JS, [
 
         // Set the custom field values
         $element->setFieldValuesFromRequest($this->_fieldsLocation);
-
-        // Mark additional fields as dirty?
-        if (!empty($this->_dirtyFields)) {
-            $element->setDirtyFields($this->_dirtyFields);
-        }
     }
 
     /**