+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added live conditional field support to inline-editable Matrix blocks. ([#14223](https://github.com/craftcms/cms/pull/14223))
 - Inline-editable Matrix blocks have been redesigned to be visually lighter. ([#14187](https://github.com/craftcms/cms/pull/14187))
 - Matrix fields set to the inline-editable blocks view mode no longer show inline entry-creation buttons unless there’s a single entry type. ([#14187](https://github.com/craftcms/cms/pull/14187))
+- Matrix fields now require the owner element to be saved before they can be edited, even when the inline-editable blocks view mode is selected.
 - Improved the accessibility of Matrix fields with the “inline-editable blocks” view mode. ([#14187](https://github.com/craftcms/cms/pull/14187))
 - Added the “Color” entry type setting. ([#14187](https://github.com/craftcms/cms/pull/14187))
 - Entry chips, cards, and blocks are now tinted according to their entry type’s color. ([#14187](https://github.com/craftcms/cms/pull/14187))
@@ -30,6 +31,7 @@
 - Fixed a bug where fields’ Name, Handle, and Instructions placeholder values within field layouts were getting set to the current overridden values.
 - Fixed a bug where nested Matrix entries could get deleted when editing multiple of them within the same parent Matrix entry.
 - Fixed a bug where empty Dropdown fields were getting treated as dirty when unchanged.
+- Fixed a bug where Matrix fields in element index or cards view weren’t working properly when nested within an inline-editable Matrix block.
 
 ## 5.0.0-alpha.8 - 2024-01-23